KeyStore: Block key attestation for Google Play Services
In order to enforce SafetyNet security, Google Play Services is now
using hardware attestation for ctsProfile validation in all cases, even
when basic attestation is selected. The SafetyNet API response from GMS
will report that basic attestation was used, but under the hood,
hardware attestation is always used regardless of the reported state.
This results in SafetyNet failing to pass due to TrustZone reporting an
unlocked bootloader (and a partially invalidated root of trust) in the
key attestation result.

We can still take advantage of the fact that this usage of hardware
attestation is opportunistic - that is, it falls back to basic
attestation if key attestation fails to run - and prevent GMS from using
key attestation at the framework level. This causes it to gracefully
fall back to basic attestation and pass SafetyNet with an unlocked
bootloader.

Key attestation is still available for other apps, as there are valid
uses for it that do not involve SafetyNet.

[ghostrider-reborn: updated for android 12 from https://github.com/kdrag0n/safetynet-fix/blob/master/java_module/app/src/main/java/dev/kdrag0n/safetynetriru/proxy/ProxyKeyStoreSpi.kt]
Change-Id: Ie85956fb0d0bb37fc4b1e9bdbfa2313e233a40c0
Co-authored-by: Adithya R <gh0strider.2k18.reborn@gmail.com>

Change-Id: Ibbd21df5e08ee9d451bb4cd122cc54f86b6e4846